+++ tanjun/dependencies/limiters.py
@@ -40,6 +40,8 @@ __all__: list[str] = [
     "CooldownPreExecution",
     "InMemoryConcurrencyLimiter",
     "InMemoryCooldownManager",
+    "add_concurrency_limit",
+    "add_cooldown",
     "with_concurrency_limit",
     "with_cooldown",
 ]
@@ -70,7 +72,6 @@ if typing.TYPE_CHECKING:
     from typing_extensions import Self
 
     _CommandT = typing.TypeVar("_CommandT", bound=tanjun.ExecutableCommand[typing.Any])
-    _OtherCommandT = typing.TypeVar("_OtherCommandT", bound=tanjun.ExecutableCommand[typing.Any])
     _InnerResourceSig = collections.Callable[[], "_InnerResourceT"]
 
 
@@ -766,27 +767,66 @@ def with_cooldown(
     Returns
     -------
     collections.abc.Callable[[tanjun.abc.ExecutableCommand], tanjun.abc.ExecutableCommand]
-        A decorator that adds a [CooldownPreExecution][tanjun.dependencies.CooldownPreExecution]
-        hook to the command.
+        A decorator which adds the relevant cooldown hooks.
+    """
+    return lambda command: _internal.apply_to_wrapped(
+        command,
+        lambda c: add_cooldown(c, bucket_id, error=error, error_message=error_message, owners_exempt=owners_exempt),
+        command,
+        follow_wrapped=follow_wrapped,
+    )
+
+
+def add_cooldown(
+    command: tanjun.ExecutableCommand[typing.Any],
+    bucket_id: str,
+    /,
+    *,
+    error: typing.Optional[collections.Callable[[str, datetime.datetime], Exception]] = None,
+    error_message: typing.Union[
+        str, collections.Mapping[str, str]
+    ] = "This command is currently in cooldown. Try again {cooldown}.",
+    owners_exempt: bool = True,
+) -> None:
+    """Add a pre-execution hook used to manage a command's cooldown.
+
+    !!! warning
+        Cooldowns will only work if there's a setup injected
+        [AbstractCooldownManager][tanjun.dependencies.InMemoryCooldownManager] dependency with
+        [InMemoryCooldownManager][tanjun.dependencies.InMemoryCooldownManager]
+        being usable as a standard in-memory cooldown manager.
+
+    Parameters
+    ----------
+    command
+        The command to add a cooldown to.
+    bucket_id
+        The cooldown bucket's ID.
+    error
+        Callback used to create a custom error to raise if the check fails.
+
+        This should two arguments one of type [str][] and [datetime.datetime][]
+        where the first is the limiting bucket's ID and the second is when said
+        bucket can be used again.
+
+        This takes priority over `error_message`.
+    error_message
+        The error message to send in response as a command error if the check fails.
+
+        This supports [localisation][] and uses the check name
+        `"tanjun.cooldown"` for global overrides.
+    owners_exempt
+        Whether owners should be exempt from the cooldown.
     """
     pre_execution = CooldownPreExecution(
         bucket_id, error=error, error_message=error_message, owners_exempt=owners_exempt
     )
+    hooks_ = command.hooks
+    if not hooks_:
+        hooks_ = hooks.AnyHooks()
+        command.set_hooks(hooks_)
 
-    def decorator(command: _OtherCommandT, /, *, _recursing: bool = False) -> _OtherCommandT:
-        hooks_ = command.hooks
-        if not hooks_:
-            hooks_ = hooks.AnyHooks()
-            command.set_hooks(hooks_)
-
-        hooks_.add_pre_execution(pre_execution)
-        if follow_wrapped and not _recursing:
-            for wrapped in _internal.collect_wrapped(command):
-                decorator(wrapped, _recursing=True)
-
-        return command
-
-    return decorator
+    hooks_.add_pre_execution(pre_execution)
 
 
 class _ConcurrencyLimit:
@@ -996,14 +1036,7 @@ class InMemoryConcurrencyLimiter(AbstractConcurrencyLimiter):
 
 
 class ConcurrencyPreExecution:
-    """Pre-execution hook used to acquire a bucket concurrency limiter.
-
-    !!! note
-        For a concurrency limiter to work properly, both
-        [ConcurrencyPreExecution][tanjun.dependencies.ConcurrencyPreExecution]
-        and [ConcurrencyPostExecution][tanjun.dependencies.ConcurrencyPostExecution]
-        hooks must be registered for a command scope.
-    """
+    """Pre-execution hook used to acquire a bucket concurrency limiter."""
 
     __slots__ = ("_bucket_id", "_error", "_error_message", "__weakref__")
 
@@ -1057,14 +1090,7 @@ class ConcurrencyPreExecution:
 
 
 class ConcurrencyPostExecution:
-    """Post-execution hook used to release a bucket concurrency limiter.
-
-    !!! note
-        For a concurrency limiter to work properly, both
-        [ConcurrencyPreExecution][tanjun.dependencies.ConcurrencyPreExecution]
-        and [ConcurrencyPostExecution][tanjun.dependencies.ConcurrencyPostExecution]
-        hooks must be registered for a command scope.
-    """
+    """Post-execution hook used to release a bucket concurrency limiter."""
 
     __slots__ = ("_bucket_id", "__weakref__")
 
@@ -1123,22 +1149,59 @@ def with_concurrency_limit(
     Returns
     -------
     collections.abc.Callable[[tanjun.abc.ExecutableCommand], tanjun.abc.ExecutableCommand]
-        A decorator that adds the concurrency limiter hooks to a command.
+        A decorator which adds the concurrency limiter hooks to a command.
+    """
+    return lambda command: _internal.apply_to_wrapped(
+        command,
+        lambda c: add_concurrency_limit(c, bucket_id, error=error, error_message=error_message),
+        command,
+        follow_wrapped=follow_wrapped,
+    )
+
+
+def add_concurrency_limit(
+    command: tanjun.ExecutableCommand[typing.Any],
+    bucket_id: str,
+    /,
+    *,
+    error: typing.Optional[collections.Callable[[str], Exception]] = None,
+    error_message: typing.Union[
+        str, collections.Mapping[str, str]
+    ] = "This resource is currently busy; please try again later.",
+) -> None:
+    """Add the hooks used to manage a command's concurrency limit.
+
+    !!! warning
+        Concurrency limiters will only work if there's a setup injected
+        [AbstractConcurrencyLimiter][tanjun.dependencies.AbstractConcurrencyLimiter] dependency with
+        [InMemoryConcurrencyLimiter][tanjun.dependencies.InMemoryConcurrencyLimiter]
+        being usable as a standard in-memory concurrency manager.
+
+    Parameters
+    ----------
+    command
+        The command to add the concurrency limit to.
+    bucket_id
+        The concurrency limit bucket's ID.
+    error
+        Callback used to create a custom error to raise if the check fails.
+
+        This should two one [str][] argument which is the limiting bucket's ID.
+
+        This takes priority over `error_message`.
+    error_message
+        The error message to send in response as a command error if this fails
+        to acquire the concurrency limit.
+
+        This supports [localisation][] and uses the check name
+        `"tanjun.concurrency"` for global overrides.
     """
     pre_execution = ConcurrencyPreExecution(bucket_id, error=error, error_message=error_message)
     post_execution = ConcurrencyPostExecution(bucket_id)
 
-    def decorator(command: _OtherCommandT, /, *, _recursing: bool = False) -> _OtherCommandT:
-        hooks_ = command.hooks
-        if not hooks_:
-            hooks_ = hooks.AnyHooks()
-            command.set_hooks(hooks_)
+    hooks_ = command.hooks
+    if not hooks_:
+        hooks_ = hooks.AnyHooks()
+        command.set_hooks(hooks_)
 
-        hooks_.add_pre_execution(pre_execution).add_post_execution(post_execution)
-        if follow_wrapped and not _recursing:
-            for wrapped in _internal.collect_wrapped(command):
-                decorator(wrapped, _recursing=True)
-
-        return command
-
-    return decorator
+    hooks_.add_pre_execution(pre_execution).add_post_execution(post_execution)